allow all build contexts to use noOpAuditor
